+++ CHANGELOG.md
@@ -5,9 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.8.3] - 2026-03-15
+## [0.8.4] - 2026-03-21
+### Added
+- 32 kHz mode in mod config
+- Custom audio debug event registration (will be displayed in debug HTML)
 ### Improved
+- Debug HTML design and features
+
+## [0.8.3] - 2026-03-15
+### Fixed
 - Heap: RSP cache invalidation
+- Mayors theme mono playback
 ### Added
 - HTML debug panel. Disabled by default; must be enabled before game start (http://127.0.0.1:18480/audio-debug.html)